refactor(7091): separate escrow and fixture responsibilities
Move escrow persistence and response models into focused modules and split shared task and escrow request fixtures by responsibility. This restores repository size limits without changing the verified lifecycle behavior.\n\nConstraint: Touched files and functions must satisfy repository size limits\nConfidence: high\nScope-risk: narrow\nReversibility: clean\nDirective: Keep escrow agreement fields colocated with escrow-specific models\nTested: strict kamn-node clippy; escrow contract cluster (35 passed, 1 ignored)\nNot-tested: Full workspace make check
