+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/task_escrow_persistence_contract_tests/support/request_support.rs
@@ -1,241 +1,15 @@
 use super::super::super::*;
-use crate::service_api_endpoint::{
-    ServiceApiAgentGetBody, ServiceApiSnapshot, ServiceApiTaskCreateBody,
+use crate::service_api_endpoint::ServiceApiSnapshot;
+
+#[path = "request_support/escrow_support.rs"]
+mod escrow_support;
+#[path = "request_support/task_support.rs"]
+mod task_support;
+
+pub(crate) use escrow_support::{fund_escrow, release_escrow};
+pub(crate) use task_support::{
+    accept_task, complete_task, create_task, query_task, register_agent_profile,
 };
-
-pub(crate) fn register_agent_profile(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    payload: &str,
-) -> ServiceApiAgentGetBody {
-    let response = raw_signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: "/v1/agents/register",
-            caller_did,
-            nonce,
-            body: payload,
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 201 Created"));
-    parse_service_api_payload(extract_http_response_body(response.as_str()))
-        .expect("agent registration payload should deserialize")
-}
-
-pub(crate) fn create_task(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    payload: &str,
-) -> ServiceApiTaskCreateBody {
-    let canonical_payload = canonical_task_payload(snapshot, bind_addr, caller_did, nonce, payload);
-    let response = signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: "/v1/tasks/create",
-            caller_did,
-            nonce,
-            body: canonical_payload.as_str(),
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 201 Created"));
-    parse_service_api_payload(extract_http_response_body(response.as_str()))
-        .expect("task create payload should deserialize")
-}
-
-fn canonical_task_payload(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    payload: &str,
-) -> String {
-    if payload.contains("\"provider_did\"") {
-        return payload.to_owned();
-    }
-    let provider = register_agent_profile(
-        snapshot,
-        bind_addr,
-        caller_did,
-        nonce - 1,
-        r#"{"agent_type":"provider","model_family":"test","capabilities":["tasks"]}"#,
-    );
-    serde_json::json!({
-        "provider_did": provider.did,
-        "transaction_id": format!("transaction-{nonce}"),
-        "terms_digest": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-        "idempotency_key": format!("create-{nonce}"),
-        "description": payload,
-    })
-    .to_string()
-}
-
-pub(crate) fn accept_task(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    task_id: &str,
-) {
-    let response = raw_signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: format!("/v1/tasks/{task_id}/accept").as_str(),
-            caller_did,
-            nonce,
-            body: r#"{"idempotency_key":"accept-task"}"#,
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 200 OK"));
-}
-
-pub(crate) fn complete_task(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    task_id: &str,
-) {
-    let path = format!("/v1/tasks/{task_id}/complete");
-    let response = raw_signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: path.as_str(),
-            caller_did,
-            nonce,
-            body: r#"{"idempotency_key":"complete-for-escrow","completion_evidence_digest":"bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"}"#,
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 200 OK"), "{response}");
-}
-
-pub(crate) fn query_task(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    task_id: &str,
-) -> Value {
-    let response = raw_signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "GET",
-            path: format!("/v1/tasks/{task_id}").as_str(),
-            caller_did,
-            nonce,
-            body: "",
-            extra_headers: &[],
-        },
-    );
-    assert!(
-        response.contains("HTTP/1.1 200 OK"),
-        "escrow release response should be 200 OK, got:\n{response}"
-    );
-    parse_service_api_payload(extract_http_response_body(response.as_str()))
-        .expect("task query payload should deserialize")
-}
-
-pub(crate) fn fund_escrow(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    payload: &str,
-) -> Value {
-    let canonical_payload = canonical_escrow_payload(caller_did, nonce, payload);
-    let response = signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: "/v1/escrow/fund",
-            caller_did,
-            nonce,
-            body: canonical_payload.as_str(),
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 200 OK"));
-    parse_service_api_payload(extract_http_response_body(response.as_str()))
-        .expect("escrow fund payload should deserialize")
-}
-
-fn canonical_escrow_payload(caller_did: &str, nonce: u64, payload: &str) -> String {
-    if payload.contains("\"transaction_id\"") {
-        return payload.to_owned();
-    }
-    let input: Value = serde_json::from_str(payload).expect("legacy escrow fixture should parse");
-    let task_id = input["task_id"]
-        .as_str()
-        .expect("task id should be present");
-    let state_file = std::env::var("KAMN_SERVICE_API_STATE_FILE").expect("state file env");
-    let state = super::env_support::read_state_json(std::path::Path::new(&state_file));
-    let task = &state["tasks"][task_id];
-    let amount = input["amount_lamports"]
-        .as_u64()
-        .or_else(|| input["amount"].as_u64())
-        .unwrap_or(1);
-    serde_json::json!({
-        "task_id": task_id,
-        "transaction_id": task["transaction_id"],
-        "beneficiary_did": task["provider_did"],
-        "amount_lamports": amount,
-        "network": "solana-devnet",
-        "terms_digest": task["terms_digest"],
-        "release_authority_did": test_service_api_sender_did(caller_did),
-        "release_policy": "task-completed",
-        "idempotency_key": format!("escrow-fund-{nonce}"),
-    })
-    .to_string()
-}
-
-pub(crate) fn release_escrow(
-    snapshot: &ServiceApiSnapshot,
-    bind_addr: &str,
-    caller_did: &str,
-    nonce: u64,
-    escrow_id: &str,
-) -> Value {
-    let body = format!(r#"{{"idempotency_key":"escrow-release-{nonce}"}}"#);
-    let response = signed_request(
-        snapshot,
-        bind_addr,
-        SignedRequest {
-            max_requests: 1,
-            method: "POST",
-            path: format!("/v1/escrow/{escrow_id}/release").as_str(),
-            caller_did,
-            nonce,
-            body: body.as_str(),
-            extra_headers: &[],
-        },
-    );
-    assert!(response.contains("HTTP/1.1 200 OK"));
-    parse_service_api_payload(extract_http_response_body(response.as_str()))
-        .expect("escrow release payload should deserialize")
-}
 
 pub(crate) struct SignedRequest<'a> {
     pub(crate) max_requests: usize,
